requirements.txt
Django==1.3.1
Fabric==1.4.0
PIL==1.1.7
South==0.7.3
-e git://github.com/django-debug-toolbar/django-debug-toolbar.git@de515e2882f1e141865a27292bcda4d394ce601d#egg=django_debug_toolbar-0.9.4-py2.6-dev
django-localeurl==1.5
django-modeltranslation==0.3.3
gunicorn==0.14.2
pycrypto==2.5
sorl-thumbnail==11.12
ssh==1.7.13
wsgiref==0.1.2
